fix: avoid overwrite to Nginx and wordpress configurations on container start

+++ entrypoint.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-cp /usr/src/nginx-defaults/wordpress.conf.include /etc/nginx/conf.d/;
-cp /usr/src/nginx-defaults/default.conf /etc/nginx/conf.d/;
 
 # if [ "$SSL_ENABLED" = "true" ]; then 
 #     cp /usr/src/nginx-defaults/options-ssl-nginx.conf /var/ssl;
@@ -12,6 +10,7 @@ sed -i "s/server_name localhost;/server_name $SERVER_NAME;/" /etc/nginx/conf.d/d
 
 if [ "$SSL_ENABLED" = "true" ]; then
     certbot --nginx -d $SERVER_NAME --non-interactive --agree-tos --register-unsafely-without-email -m admin@$SERVER_NAME
+    # cerrtbot started nginx but we need to stop it for now. Later we will start it in the foreground.
     service nginx stop
 fi
 
@@ -24,15 +23,17 @@ if [ ! -f /var/www/html/index.php ]; then
     rm -rf /var/www/wordpress
 fi
 
-# Copy wp-config-sample.php to wp-config.php and update database configuration
-cp /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
-sed -i "s/database_name_here/${WP_DB_NAME}/" /var/www/html/wp-config.php
-sed -i "s/username_here/${WP_DB_USER}/" /var/www/html/wp-config.php
-sed -i "s/password_here/${WP_DB_PASSWORD}/" /var/www/html/wp-config.php
-sed -i "s/localhost/${WP_DB_HOST}/" /var/www/html/wp-config.php
+# If wp-config.php does not exist and wp-config-sample.php exists, copy wp-config-sample.php to wp-config.php and update database configuration
+if [ ! -f /var/www/html//wp-config.php ] && [ -f /var/www/html/wp-config-sample.php ]; then
+    cp /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+    sed -i "s/database_name_here/${WP_DB_NAME}/" /var/www/html/wp-config.php
+    sed -i "s/username_here/${WP_DB_USER}/" /var/www/html/wp-config.php
+    sed -i "s/password_here/${WP_DB_PASSWORD}/" /var/www/html/wp-config.php
+    sed -i "s/localhost/${WP_DB_HOST}/" /var/www/html/wp-config.php
 
-# Change owner of the web folder
-chown -R www-data /var/www/html
+    # Change owner of the web folder to make sure proper permissions for nginx
+    chown -R www-data /var/www/html
+fi
 
 # Start PHP-FPM in the background
 php-fpm &
